misc(refactor): refactor init
