--- .yo-rc.json
@@ -1,54 +1,0 @@
-{
-  "webpack-addons-demo": {
-    "configuration": {
-      "dev": {
-        "webpackOptions": {}
-      }
-    }
-  },
-  "webpack-cli": {
-    "configuration": {
-      "config": {
-        "webpackOptions": {
-          "module": {
-            "rules": [
-              {
-                "test": {},
-                "use": [
-                  {
-                    "loader": "'style-loader'"
-                  },
-                  {
-                    "loader": "'css-loader'"
-                  },
-                  {
-                    "loader": "'sass-loader'"
-                  }
-                ]
-              }
-            ]
-          },
-          "plugins": [
-            "new UglifyJSPlugin()"
-          ],
-          "entry": {
-            "a": "\"d.js\""
-          },
-          "output": {
-            "filename": "'[name].[chunkhash].js'",
-            "chunkFilename": "'[name].[chunkhash].js'",
-            "path": "path.resolve(__dirname, 'dist')"
-          }
-        },
-        "topScope": [
-          "const webpack = require('webpack')",
-          "const path = require('path')",
-          "/*\n * We've enabled UglifyJSPlugin for you! This minifies your app\n * in order to load faster and run less javascript.\n *\n * https://github.com/webpack-contrib/uglifyjs-webpack-plugin\n *\n */",
-          "const UglifyJSPlugin = require('uglifyjs-webpack-plugin');",
-          "\n",
-          "/*\n * We've enabled commonsChunkPlugin for you. This allows your app to\n * load faster and it splits the modules you provided as entries across\n * different bundles!\n *\n * https://webpack.js.org/plugins/commons-chunk-plugin/\n *\n */"
-        ]
-      }
-    }
-  }
-}
