+++ lib/generators/init-generator.js
@@ -387,7 +387,6 @@ module.exports = class InitGenerator extends Generator {
 				done();
 			});
 	}
-	/*
 	installPlugins() {
 		let asyncNamePrompt = this.async();
 		let defaultName = this.isProd ? "prod" : "config";
@@ -411,8 +410,7 @@ module.exports = class InitGenerator extends Generator {
 				});
 			});
 	}
-	*/
 	writing() {
-		this.config.set('configuration', this.configuration);
+		this.config.set("configuration", this.configuration);
 	}
 };
