+++ lib/init/index.js
@@ -5,35 +5,35 @@ const j = require("jscodeshift");
 const chalk = require("chalk");
 const pEachSeries = require("p-each-series");
 
-const runPrettier = require("../../utils/run-prettier");
+const runPrettier = require("../utils/run-prettier");
 
-const entryTransform = require("./entry/entry");
-const outputTransform = require("./output/output");
-const contextTransform = require("./context/context");
-const resolveTransform = require("./resolve/resolve");
-const devtoolTransform = require("./devtool/devtool");
-const targetTransform = require("./target/target");
-const watchTransform = require("./watch/watch");
-const watchOptionsTransform = require("./watch/watchOptions");
-const externalsTransform = require("./externals/externals");
-const nodeTransform = require("./node/node");
-const performanceTransform = require("./performance/performance");
-const statsTransform = require("./stats/stats");
-const amdTransform = require("./other/amd");
-const bailTransform = require("./other/bail");
-const cacheTransform = require("./other/cache");
-const profileTransform = require("./other/profile");
-const mergeTransform = require("./other/merge");
-const parallelismTransform = require("./other/parallelism");
-const recordsInputPathTransform = require("./other/recordsInputPath");
-const recordsOutputPathTransform = require("./other/recordsOutputPath");
-const recordsPathTransform = require("./other/recordsPath");
-const moduleTransform = require("./module/module");
-const pluginsTransform = require("./plugins/plugins");
-const topScopeTransform = require("./top-scope/top-scope");
-const devServerTransform = require("./devServer/devServer");
-const modeTransform = require("./mode/mode");
-const resolveLoaderTransform = require("./resolveLoader/resolveLoader");
+const entryTransform = require("./transformations/entry/entry");
+const outputTransform = require("./transformations/output/output");
+const contextTransform = require("./transformations/context/context");
+const resolveTransform = require("./transformations/resolve/resolve");
+const devtoolTransform = require("./transformations/devtool/devtool");
+const targetTransform = require("./transformations/target/target");
+const watchTransform = require("./transformations/watch/watch");
+const watchOptionsTransform = require("./transformations/watch/watchOptions");
+const externalsTransform = require("./transformations/externals/externals");
+const nodeTransform = require("./transformations/node/node");
+const performanceTransform = require("./transformations/performance/performance");
+const statsTransform = require("./transformations/stats/stats");
+const amdTransform = require("./transformations/other/amd");
+const bailTransform = require("./transformations/other/bail");
+const cacheTransform = require("./transformations/other/cache");
+const profileTransform = require("./transformations/other/profile");
+const mergeTransform = require("./transformations/other/merge");
+const parallelismTransform = require("./transformations/other/parallelism");
+const recordsInputPathTransform = require("./transformations/other/recordsInputPath");
+const recordsOutputPathTransform = require("./transformations/other/recordsOutputPath");
+const recordsPathTransform = require("./transformations/other/recordsPath");
+const moduleTransform = require("./transformations/module/module");
+const pluginsTransform = require("./transformations/plugins/plugins");
+const topScopeTransform = require("./transformations/top-scope/top-scope");
+const devServerTransform = require("./transformations/devServer/devServer");
+const modeTransform = require("./transformations/mode/mode");
+const resolveLoaderTransform = require("./transformations/resolveLoader/resolveLoader");
 
 /**
  *
