+++ lib/utils/modify-config-helper.js
@@ -4,7 +4,7 @@ const fs = require("fs");
 const path = require("path");
 const chalk = require("chalk");
 const yeoman = require("yeoman-environment");
-const runTransform = require("../init/transformations/index");
+const runTransform = require("../init/index");
 const Generator = require("yeoman-generator");
 
 /**
@@ -26,8 +26,8 @@ module.exports = function modifyHelperUtil(action, generator, packages) {
 	const env = yeoman.createEnv("webpack", null);
 	const generatorName = `webpack-${action}-generator`;
 
-	if(!generator) {
-		generator = class extends  Generator {
+	if (!generator) {
+		generator = class extends Generator {
 			initializing() {
 				packages.forEach(pkgPath => {
 					return this.composeWith(require.resolve(pkgPath));
@@ -40,21 +40,26 @@ module.exports = function modifyHelperUtil(action, generator, packages) {
 	env.run(generatorName).on("end", () => {
 		let configModule;
 		try {
-			const configPath = path.resolve(process.cwd(), '.yo-rc.json')
+			const configPath = path.resolve(process.cwd(), ".yo-rc.json");
 			configModule = require(configPath);
-				// Change structure of the config to be transformed
-				let tmpConfig = {};
-				Object.keys(configModule).forEach( prop => {
-					const Configs = Object.keys(configModule[prop].configuration)
-					Configs.forEach ( (config) => {
-						tmpConfig[config] = configModule[prop].configuration[config]
-					})
+			// Change structure of the config to be transformed
+			let tmpConfig = {};
+			Object.keys(configModule).forEach(prop => {
+				const Configs = Object.keys(configModule[prop].configuration);
+				Configs.forEach(config => {
+					tmpConfig[config] = configModule[prop].configuration[config];
 				});
-				configModule = tmpConfig;
-			
+			});
+			configModule = tmpConfig;
 		} catch (err) {
-			console.error(chalk.red("\nCould not find a yeoman configuration file.\n"))
-			console.error(chalk.red("\nPlease make sure to use 'this.config.set('configuration', this.configuration);' at the end of the generator.\n"))
+			console.error(
+				chalk.red("\nCould not find a yeoman configuration file.\n")
+			);
+			console.error(
+				chalk.red(
+					"\nPlease make sure to use 'this.config.set('configuration', this.configuration);' at the end of the generator.\n"
+				)
+			);
 			Error.stackTraceLimit = 0;
 			process.exitCode = -1;
 		}
