+++ lib/utils/resolve-packages.js
@@ -45,7 +45,7 @@ function resolvePackages(pkg) {
 
 	function invokeGeneratorIfReady() {
 		if (packageLocations.length === pkg.length)
-			return modifyConfigHelper("init", null, packageLocations);;
+			return modifyConfigHelper("init", null, packageLocations);
 	}
 
 	pkg.forEach(addon => {
