--- webpack.config.js
@@ -1,55 +1,0 @@
-const webpack = require('webpack');
-const path = require('path');
-
-/*
- * We've enabled UglifyJSPlugin for you! This minifies your app
- * in order to load faster and run less javascript.
- *
- * https://github.com/webpack-contrib/uglifyjs-webpack-plugin
- *
- */
-
-const UglifyJSPlugin = require('uglifyjs-webpack-plugin');
-
-/*
- * We've enabled commonsChunkPlugin for you. This allows your app to
- * load faster and it splits the modules you provided as entries across
- * different bundles!
- *
- * https://webpack.js.org/plugins/commons-chunk-plugin/
- *
- */
-
-module.exports = {
-	entry: {
-		a: d.js
-	},
-
-	output: {
-		filename: '[name].[chunkhash].js',
-		chunkFilename: '[name].[chunkhash].js',
-		path: path.resolve(__dirname, 'dist')
-	},
-
-	module: {
-		rules: [
-			{
-				test: {},
-
-				use: [
-					{
-						loader: 'style-loader'
-					},
-					{
-						loader: 'css-loader'
-					},
-					{
-						loader: 'sass-loader'
-					}
-				]
-			}
-		]
-	},
-
-	plugins: [new UglifyJSPlugin()]
-};
